(feat) basic SQL queries written for client-side interactions, database instructions written
