+++ server/middleware.js
@@ -1,6 +1,7 @@
 var morgan = require('morgan');
 var bodyParser = require('body-parser');
-var databaseController = require('./db/databaseController.js');
+var databaseController = require('./db/Mongo/databaseController.js');
+var SQLController = require('./db/SQL/SQLController.js')
 
 module.exports = function(app, express){
